requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
appdirs==1.4.4            # via virtualenv
aspy.yaml==1.3.0          # via pre-commit
attrs==19.3.0             # via pytest
cfgv==2.0.1               # via pre-commit
coverage==5.2             # via pytest-cov
distlib==0.3.1            # via virtualenv
filelock==3.0.12          # via virtualenv
flake8==3.8.3             # via -r requirements.in
identify==1.4.21          # via pre-commit
importlib-metadata==1.7.0  # via flake8, pluggy, pre-commit, pytest, virtualenv
importlib-resources==3.0.0  # via pre-commit, virtualenv
isort==4.3.21             # via -r requirements.in
mccabe==0.6.1             # via flake8
more-itertools==8.4.0     # via pytest
nodeenv==1.4.0            # via pre-commit
numpy==1.18.5             # via pandas
packaging==20.4           # via pytest
pandas==0.23.0            # via -r requirements.in
pathlib2==2.3.5           # via pytest
pluggy==0.13.1            # via pytest
pre-commit==1.21.0        # via -r requirements.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements.in
pytest==5.4.3             # via -r requirements.in, pytest-cov
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via pandas
pyyaml==5.3.1             # via aspy.yaml, pre-commit
six==1.15.0               # via cfgv, packaging, pathlib2, pre-commit, python-dateutil, virtualenv
toml==0.10.1              # via pre-commit
virtualenv==20.0.25       # via pre-commit
wcwidth==0.2.5            # via pytest
zipp==1.2.0               # via importlib-metadata, importlib-resources
